Get rid of PVS-studio warnings

+++ CPP/Async.cpp
@@ -53,14 +53,14 @@ vector<string> readData(ifstream& file) {
 
 //! vec is not const, because latter it's iterators will be used to modify it's
 //! elements -- see wordCounter.
-vector<vector<string>::iterator> SplitVector(vector<string>& vec, unsigned n) {
+vector<vector<string>::iterator> SplitVector(vector<string>& vec, size_t n) {
 
     vector<vector<string>::iterator> outVec;
     auto part_length = vec.size() / n;
     auto sum = vec.begin();
     outVec.push_back(vec.begin());
 
-    for(int i = 0; i<n-1;++i){
+    for(size_t i = 0; i<n-1;++i){
         advance(sum, part_length);
         outVec.push_back(sum);
         // cout<<outVec[i]<<endl;
@@ -77,7 +77,7 @@ int main(int argc, char *argv[]) {
     string infile    = config["infile"];
     string out_by_a  = config["out_by_a"];
     string out_by_n  = config["out_by_n"];
-    int    threads_n = str_to_val<unsigned>(config["threads"]);
+    size_t threads_n = str_to_val<size_t>(config["threads"]);
 
     string etalon_a_file  = config["etalon_a_file"];
